refactor : 테스트 코드 디렉토리 위치 수정
