+++ src/test/java/nextstep/laddergame/domain/LineTest.java
@@ -1,9 +1,8 @@
-package nextstep.laddergame;
+package nextstep.laddergame.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.List;
-import nextstep.laddergame.domain.Line;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
