+++ src/test/java/nextstep/laddergame/domain/ParticipantTest.java
@@ -1,7 +1,6 @@
-package nextstep.laddergame;
+package nextstep.laddergame.domain;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import nextstep.laddergame.domain.Participant;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
